rename files, check program id NaN
