+++ pages/api/class/index.ts
@@ -3,7 +3,7 @@ import { ResponseUtil } from "@utils/responseUtil";
 import { getClasses, createClass } from "@database/class";
 import { ClassInput } from "@models/Class";
 import { validateClassData } from "@utils/validation/class";
-import { getClassInfoWithProgramId } from "@database/programcardinfo";
+import { getClassInfoWithProgramId } from "@database/program-card-info";
 
 /**
  * handle controls the request made to the class resource
@@ -16,12 +16,24 @@ export default async function handle(
 ): Promise<void> {
     switch (req.method) {
         case "GET": {
-            const { id } = req.query;
+            const { id: programId } = req.query;
 
-            const classes = id
-                ? await getClassInfoWithProgramId(id as string)
-                : await getClasses();
-            ResponseUtil.returnOK(res, classes);
+            if (!programId) {
+                const classes = await getClasses();
+                ResponseUtil.returnOK(res, classes);
+            } else {
+                const programIdNumber = parseInt(programId as string, 10);
+                if (isNaN(programIdNumber)) {
+                    return ResponseUtil.returnBadRequest(
+                        res,
+                        "programId should be passed in as numbers",
+                    );
+                }
+                const classes = await getClassInfoWithProgramId(
+                    programId as string,
+                );
+                ResponseUtil.returnOK(res, classes);
+            }
             break;
         }
         case "POST": {
